requirements.txt
django
djangorestframework
django-filter
markdown
celery
django-celery-beat
gevent
requests
sqlalchemy

